added some reactstrap styling
