+++ crypto-price/package.json
@@ -7,11 +7,14 @@
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
     "axios": "^0.19.2",
+    "bootstrap": "^4.4.1",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-loader-spinner": "^3.1.5",
     "react-redux": "^7.1.3",
     "react-scripts": "3.3.0",
+    "react-strap": "^0.0.1",
+    "reactstrap": "^8.4.0",
     "redux": "^4.0.5",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0"
