+++ crypto-price/src/index.js
@@ -6,9 +6,10 @@ import { Provider } from 'react-redux';
 import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import logger from 'redux-logger';
+import 'bootstrap/dist/css/bootstrap.min.css';
 
 import { reducer } from './reducers';
-import App from './App';
+import App from './components/App';
 
 const store = createStore(reducer, applyMiddleware(thunk, logger));
